Convert mask to float in gpu device mode
Ensure masks are converted to float32 when moved to GPU, which is required for CUDA's max_pool2d operation used in mask expansion and fill holes.

+++ inpaint_cropandstitch.py
@@ -1287,8 +1287,8 @@ class InpaintCropImproved:
         if device_mode == "gpu (much faster)":
             device = comfy.model_management.get_torch_device()
             image = image.to(device)
-            if mask is not None: mask = mask.to(device)
-            if optional_context_mask is not None: optional_context_mask = optional_context_mask.to(device)
+            if mask is not None: mask = mask.to(device).float()
+            if optional_context_mask is not None: optional_context_mask = optional_context_mask.to(device).float()
             processor = GPUProcessorLogic()
         else:
             processor = CPUProcessorLogic()
